+++ pyjsonrpc/rpcerror.py
@@ -10,10 +10,12 @@ class JsonRpcError(RuntimeError):
     message = None
     data = None
 
-    def __init__(self, message = None, data = None):
+    def __init__(self, message = None, data = None, code = None):
         RuntimeError.__init__(self)
         self.message = message or self.message
         self.data = data
+        self.code = code or self.code
+        assert self.code, "Error without code is not allowed."
 
 jsonrpcerrors[JsonRpcError.code] = JsonRpcError
 
@@ -22,12 +24,18 @@ class ParseError(JsonRpcError):
     code = -32700
     message = u"Invalid JSON was received by the server."
 
+    def __init__(self, message = None, data = None):
+        JsonRpcError.__init__(self, message = message, data = data)
+
 jsonrpcerrors[ParseError.code] = ParseError
 
 
 class InvalidRequest(JsonRpcError):
     code = -32600
     message = u"The JSON sent is not a valid Request object."
+
+    def __init__(self, message = None, data = None):
+        JsonRpcError.__init__(self, message = message, data = data)
 
 jsonrpcerrors[InvalidRequest.code] = InvalidRequest
 
@@ -36,12 +44,18 @@ class MethodNotFound(JsonRpcError):
     code = -32601
     message = u"The method does not exist / is not available."
 
+    def __init__(self, message = None, data = None):
+        JsonRpcError.__init__(self, message = message, data = data)
+
 jsonrpcerrors[MethodNotFound.code] = MethodNotFound
 
 
 class InvalidParams(JsonRpcError):
     code = -32602
     message = u"Invalid method parameter(s)."
+
+    def __init__(self, message = None, data = None):
+        JsonRpcError.__init__(self, message = message, data = data)
 
 jsonrpcerrors[InvalidParams.code] = InvalidParams
 
@@ -50,5 +64,9 @@ class InternalError(JsonRpcError):
     code = -32603
     message = u"Internal JSON-RPC error."
 
+    def __init__(self, message = None, data = None):
+        JsonRpcError.__init__(self, message = message, data = data)
+
 jsonrpcerrors[InternalError.code] = InternalError
+